Rename and remove electron branches when reading
Common changes with the muon branches are described in b70724fc481b675205547d1f52cfb495ea283fd8
(forgot to mention that the nMuon and nElectron types are now UInt_t instead of Int_t, and that passes*IdPOG are now booleans instead of integers)

Renamed & removed the following:
- eleMVArawSpring16GP -> mvaSpring16GP
- eleMVArawSpring16HZZ -> mvaSpring16HZZ
- eleSieie -> sieie
- eleHoE -> hoe
- eleooEmooP -> eInvMinusPInv

Changes in branch types:
- lostHits: Int_t -> UChar_t
- conversionVeto: Int_t -> Bool_t

NB! I have created additional two placeholder branches that we currently read until it becomes clear how this information will be retained in the nanoAOD format:
- eleDEta
- eleDPhi

+++ src/ElectronHistManager.cc
@@ -64,7 +64,7 @@ void ElectronHistManager::fillHistograms(const RecoElectron& electron, double ev
   fillWithOverFlow(histogram_deltaPhi_, electron.deltaPhi(), evtWeight, evtWeightErr);
   fillWithOverFlow(histogram_OoEminusOoP_, electron.OoEminusOoP(), evtWeight, evtWeightErr);
   fillWithOverFlow(histogram_nLostHits_, electron.nLostHits(), evtWeight, evtWeightErr);
-  fillWithOverFlow(histogram_passesConversionVeto_, electron.passesConversionVeto(), evtWeight, evtWeightErr);
+  fillWithOverFlow(histogram_passesConversionVeto_, static_cast<int>(electron.passesConversionVeto()), evtWeight, evtWeightErr);
   
   int abs_genPdgId = 0;
   if      ( electron.genLepton() ) abs_genPdgId = std::abs(electron.genLepton()->pdgId()); // generator level match to electron or muon
